+++ src-dbg/flow_storm/debugger/events_processor.clj
@@ -149,8 +149,8 @@
                   (seq (dbg-state/flow-exceptions flow-id)))))
      (flows-screen/goto-location bookmark-location))))
 
-(defn data-window-push-val-data-event [{:keys [dw-id val-data root? visualizer]}]
-  (data-windows/push-val dw-id val-data {:root? root?, :visualizer visualizer}))
+(defn data-window-push-val-data-event [{:keys [dw-id val-data extras]}]
+  (data-windows/push-val dw-id val-data extras))
 
 (defn data-window-update-event [{:keys [dw-id data]}]
   (data-windows/update-val dw-id data))