+++ src-inst/flow_storm/api.clj
@@ -354,10 +354,11 @@
   ([dw-id val] (data-window-push-val dw-id val nil))
   ([dw-id val stack-key] (data-window-push-val dw-id val stack-key nil))
   ([dw-id val stack-key extra]
-   (let [vdata (assoc (rt-values/extract-data-aspects val extra)
-                      :flow-storm.debugger.ui.data-windows.data-windows/dw-id dw-id
-                      :flow-storm.debugger.ui.data-windows.data-windows/stack-key stack-key)
-         extra (update extra :root? (fn [root?] (if (nil? root?) true root?)))]
+   (let [vdata (rt-values/extract-data-aspects val extra)
+         extra (assoc extra
+                      :dw-id dw-id
+                      :stack-key stack-key
+                      :root? true)]
      (rt-events/publish-event! (rt-events/make-data-window-push-val-data-event dw-id vdata extra)))))
 
 (defn data-window-val-update [dw-id new-val]